+++ web/modules/custom/ai_screening_project/src/Helper/ProjectHelper.php
@@ -23,11 +23,13 @@ use Drupal\core_event_dispatcher\Event\Entity\EntityAccessEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityBaseFieldInfoEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityDeleteEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityInsertEvent;
+use Drupal\group\Entity\GroupInterface;
 use Drupal\group\Entity\GroupRelationshipInterface;
 use Drupal\group\Entity\Storage\GroupRelationshipStorageInterface;
 use Drupal\group\Entity\Storage\GroupStorage;
 use Drupal\node\NodeInterface;
 use Drupal\node\NodeStorageInterface;
+use Drupal\preprocess_event_dispatcher\Event\NodePreprocessEvent;
 use Drupal\taxonomy\TermStorageInterface;
 use Drupal\user\UserStorageInterface;
 use Drupal\webform\WebformSubmissionStorageInterface;
@@ -265,6 +267,7 @@ class ProjectHelper extends AbstractHelper implements EventSubscriberInterface {
       EntityHookEvents::ENTITY_ACCESS => 'entityAccess',
       EntityHookEvents::ENTITY_INSERT => 'entityInsert',
       EntityHookEvents::ENTITY_DELETE => 'entityDelete',
+      NodePreprocessEvent::name('project') => 'preprocessProject',
       // @fixme I, Mikkel, cannot make this work using an event handler, so we
       // do it the old fashioned way with a hook implementation in
       // ai_screening_project.module (which see).
@@ -338,7 +341,9 @@ class ProjectHelper extends AbstractHelper implements EventSubscriberInterface {
       foreach ($projectTrackTerms as $projectTrackTerm) {
         $projectTrack = $this->projectTrackStorage
           ->create([
-            'type' => 'project_group',
+            'type' => $projectTrackTerm->id(),
+            'title' => $projectTrackTerm->getName(),
+            'description' => $projectTrackTerm->getDescription(),
             'project_track_evaluation' => '0',
             'project_id' => $entity,
           ])
@@ -401,6 +406,22 @@ class ProjectHelper extends AbstractHelper implements EventSubscriberInterface {
   }
 
   /**
+   * Preprocess node project event.
+   *
+   * @param \Drupal\preprocess_event_dispatcher\Event\NodePreprocessEvent $event
+   *   The event being performed.
+   */
+  public function preprocessProject(NodePreprocessEvent $event): void {
+    $variables = $event->getVariables();
+    $node = $variables->getEntity();
+    if ($node instanceof NodeInterface) {
+      $variables->set('projectTracks', $this->loadProjectTracks($node));
+      $variables->set('projectGroup', $this->loadProjectGroup($node));
+      $variables->set('projectMembers', $this->loadProjectGroup($node)->getRelatedEntities('group_membership'));
+    }
+  }
+
+  /**
    * Load project tracks.
    *
    * @param \Drupal\node\NodeInterface $project
@@ -417,6 +438,27 @@ class ProjectHelper extends AbstractHelper implements EventSubscriberInterface {
       ->execute();
 
     return $this->projectTrackStorage->loadMultiple($ids);
+  }
+
+  /**
+   * Load project group.
+   *
+   * @param \Drupal\node\NodeInterface $project
+   *   The project.
+   *
+   * @return \Drupal\group\Entity\GroupInterface
+   *   The group
+   */
+  public function loadProjectGroup(NodeInterface $project) : GroupInterface {
+    $relationshipIds = $this->groupRelationshipStorage->getQuery()
+      ->accessCheck(FALSE)
+      ->condition('entity_id', $project->id(), '=')
+      ->condition('type', 'project_group-group_node-project', '=')
+      ->execute();
+
+    $relationships = $this->groupRelationshipStorage->loadMultiple($relationshipIds);
+
+    return $this->groupStorage->load(reset($relationships)->getGroupId());
   }
 
 }