+++ web/themes/custom/itkdev/itkdev_project_theme/js/project-icons.js
@@ -1,3 +1,3 @@
 export const projectFaBrandIcons = ['faTwitter'];
-export const projectFaSolidIcons = ['faAngleRight', 'faCirclePlus', 'faSortDown', 'faSortUp'];
+export const projectFaSolidIcons = ['faAngleRight', 'faPencil', 'faCirclePlus', 'faSortDown', 'faSortUp'];
 export const projectFaRegularIcons = ['faAngleLeft'];